Added checks so > 32bit values are not passed through COM for non XP systems

+++ ViewShare.h
@@ -127,10 +127,10 @@ protected:
 	void LoadSharedList();
 	void RunSelected();
 
-	DWORD TotalFiles;
-	DWORD TotalSize;
-	DWORD TotalMatched;
-	DWORD TotalUploaded;
+	DWORD  TotalFiles;
+	uint64 TotalSize;
+	DWORD  TotalMatched;
+	DWORD  TotalUploaded;
 
 	CGnucleusDoc* m_pDoc;
 	CAutShare*    m_autShare;
